FEATURE: cleaner code in a single option; passes 2 of 2 "messy tests"
